moved functions, modified Hereditary and generalizedSplines
isPureT: checks if input complex is pure and top dimensional
getDim: returns dimension of affine span of a set of vertices
isHereditary: modified so that it takes as input a list of vertices and
facets.  Then it checks that maximal intersections of facets are
codimension one and if so, uses these as codim 2 faces. Still needs to
be modified (link of every face needs to be checked, not just vertices).
generalizedSplines: allow for more flexible input, including mod n, so
that it will be useful for folks studying generalized splines.